Couple of updates: telegraf line protocol, dynamic imports, metadata
This change uses the pkgutil module to import the plugins and load the
selected module. This same method is also used to list all of the
available modules.

To test run without installing the app pah has been added to the module
main using the sys path as found at runtime.

The telegraf line protocol has been added to allow this project to be
run with an exec stanza within telegraf allowing it to store metrics in
InfluxDB.

Meta has been added to the example plugin allowing additional meta data
to be returned in the result. This provides the ability to add tags or
other metadata into a given monitoring system using a simple key=value
format.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>